refactor(console-commands): fix parser conflicts between similarly named events between plugins
