+++ SuperEvents/SuperEvents.csproj
@@ -60,6 +60,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="EventFunctions\API.cs" />
+    <Compile Include="EventFunctions\EventParameterAutoCompleter.cs" />
     <Compile Include="EventFunctions\VersionChecker.cs" />
     <Compile Include="Events\CarAccident.cs" />
     <Compile Include="Events\CarFire.cs" />
